Ensure that Objptrs are aligned in Sequence objects
With `-align 4`, no padding was added to the element components of a
Sequence object.  This could result in Sequence object types of the
form:

    Sequence {components = ([Word1, Bits7],
                            Objptr (opt_144),
                            Objptr (opt_5)),
              hasIdentity = false}

This places the Objptr components at offsets 1 and 5.  For x86, such
misaligned accesses are accepted (possibly at a performance cost);
other 32-bit platforms might not accept such missaligned accesses.

However, this representation is incompatible with the mark-compact
GC (Jonker's compacting collector), which must be able to distinguish
object headers from object pointers when threading.  In particular,
the runtime uses `GC_VALID_HEADER_MASK & header`, where
`#define GC_VALID_HEADER_MASK ((GC_header)0x1)`, to determine whether
or not an object is threaded by examining its header.

When the first object pointer in a sequence like that above is
threaded, this will result in the address of that object pointer being
stored in a header field (of the object pointed-to by the object
pointer in the sequence).  But, the address of the first object
pointer is at offset 1 from the start of the sequence object and is
therefore at an address with a low bit of 1.  This means that the
threaded pointer will be mistakenly interpreted as a header and either
an assertion failure or segmentation fault.

To address the issue, padding is added to the element components of a
Sequence object to ensure that the width of an element is aligned to
the minimum of the requested global alignment (i.e., `-align <n>`) and
the maximum size of a component.  So, for instance, a sequence with
only a Word8 component will require no padding, while a sequence with
a Word8 component and an Objptr component (on a 32-bit system) will
require 24 bytes of padding.  The above Sequence object is now
represented as:

    Sequence {components = ([Word1, Bits7],
                            Bits16,
                            Bits8,
                            Objptr (opt_144),
                            Objptr (opt_5)),
              hasIdentity = false}

With `-align 8`, this was already handled by ensuring that any
sequence with an element component of width 8 bytes (e.g., an Objptr)
was padded so that the width of an element is 8-byte aligned.

This does increase the size of some sequence objects.  Technically,
only sequences with objptr components need to have padding and only
need to be 2-byte aligned.  However, for uniformity with `-align 8`
and for faster data access, aligning to the minimum of the global
alignment and the maximum size of a component is chosen.

To save space and avoid misaligned objptr data, it would be better to
perform an "sequence of tuples" to "tuple of sequences" transformation
in the SSA2 IR.