Fix bug in `ImperativeIO.getInstream`
Properly flush the "imperative" ImperativeIO buffer when realizing the
underlying stream via `getInstream`.

Failing to reset the `first` and `last` refs when realizing the
`StreamIO.instream` from an `ImperativeIO.instream` that is in
"imperative" mode allows subsequent `ImperativeIO.input*` functions to
advance the stream with respect to the "imperative" mode but not
advance the `StreamIO.instream`.  After realizing the
`StreamIO.instream` from an `ImpearativeIO.instream`, the stream
should be in "stream" mode and always advance the `StreamIO.instream`.

Fixes MLton/mlton#535