Add `-profile-tail-call-opt {true|false}` expert compile-time option
The `-profile-tail-call-opt {true|false}` controls whether or not the
SSA{,2} shrinker optimizes tail calls in the presence of profiling.

`-profile-tail-call-opt false` is expected to have a significant
performance penalty, but can improve the accuracy of exception
history.  It likely worsens the accuracy of time profiling, since the
profiled program (without tail call optimizations) will be
significantly different from the non-profiled program (with tail call
optimizations).