Add `-profile-intro-loops-opt {true|false}` expert compile-time option
The `-profile-intro-loops-opt {true|false}` controls whether or not the
SSA IntroduceLoops optimization applies in presence of profiling.

In particular, when `-profile-tail-call-opt false` but
`-profile-intro-loops-opt true`, then IntroduceLoops will recognize
self non-tail calls with eta return and handler continuations as tail
calls.  This is expected to recover some of the performance penalty of
`-profile-tail-call-opt false`, at the expense of less accurate
exception history; the exception history will have only one entry for
the recursive function, even if the exception is raised by a deeply
nested recursive (tail) call.