Fix bug in `SplitTypes` SSA optimization
The `Array_copy{Array,Vector}` primitives correspond to a `memcpy` of
elements from one sequence to another and require the element types to
be the same.  `SplitTypes` was not unifying the `TypeInfo` of the
elements of the source and destination sequences, allowing their types
to be split, leading to an internal SSA IR type checking error.

Thanks to David Wang for the bug report.